+++ History.md
@@ -1,6 +1,7 @@
 0.1.1 / 2013-?-?
 ==================
 
+  * Added multidatabase support
   * Fixed element name bug
   * Fixed problem with infinite loop made by relations.
   * Init now expects objects for relation, not id's.